Move staking test helpers to the test-utils package
