+++ packages/test-utils/Cargo.toml
@@ -8,10 +8,13 @@ authors.workspace = true
 publish.workspace = true
 
 [dependencies]
+babylon-apis     = { path = "../apis" }
 babylon-proto    = { path = "../proto" }
 babylon-bitcoin  = { path = "../bitcoin" }
 cargo_metadata   = { workspace = true }
+cosmwasm-std     = { workspace = true }
 hex              = { workspace = true }
+k256             = { workspace = true }
 prost            = { workspace = true }
 serde            = { workspace = true }
 serde_json       = { workspace = true }
